test/web-platform-tests/to-upstream/css/cssom-view/HTMLElement-offsetParent.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTMLElement.prototype.offsetParent</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-htmlelement-offsetparent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_equals(el.offsetParent, null);

}, "offsetParent should return null for an element disconnected from a document");
</script>
